+++ src/core/version.js
